Change datatype for time to match Qt.

+++ xMainMovieWidget.cpp
@@ -352,10 +352,10 @@ void xMainMovieWidget::updatePlayedMovies() {
                 auto playCount = std::get<1>(*playedMovie);
                 if (playCount > 1) {
                     movieItem->addToolTip(QString(tr("played %1 times, last time on %2")).arg(playCount).
-                            arg(QDateTime::fromMSecsSinceEpoch(static_cast<qint64>(std::get<2>(*playedMovie))).toString(Qt::DefaultLocaleLongDate)));
+                            arg(QDateTime::fromMSecsSinceEpoch(std::get<2>(*playedMovie)).toString(Qt::DefaultLocaleLongDate)));
                 } else {
                     movieItem->addToolTip(QString(tr("played once, last time on %1")).
-                            arg(QDateTime::fromMSecsSinceEpoch(static_cast<qint64>(std::get<2>(*playedMovie))).toString(Qt::DefaultLocaleLongDate)));
+                            arg(QDateTime::fromMSecsSinceEpoch(std::get<2>(*playedMovie)).toString(Qt::DefaultLocaleLongDate)));
                 }
                 // Remove element to speed up search in the next iteration.
                 playedMovies.erase(playedMovie);
@@ -371,7 +371,7 @@ void xMainMovieWidget::updatePlayedMovies() {
 }
 
 void xMainMovieWidget::updatePlayedMovie(const QString& tag, const QString& directory,
-                                         const QString& movie, int playCount, quint64 timeStamp) {
+                                         const QString& movie, int playCount, qint64 timeStamp) {
     // Only update if the database overlay is enabled.
     if (!useDatabaseMovieOverlay) {
         return;
@@ -406,10 +406,10 @@ void xMainMovieWidget::updatePlayedMovie(const QString& tag, const QString& dire
         moviePlayedItem->setIcon(":images/xplay-star.svg");
         if (playCount > 1) {
             moviePlayedItem->addToolTip(QString(tr("played %1 times, last time on %2")).arg(playCount).
-                    arg(QDateTime::fromMSecsSinceEpoch(static_cast<qint64>(timeStamp)).toString(Qt::DefaultLocaleLongDate)));
+                    arg(QDateTime::fromMSecsSinceEpoch(timeStamp).toString(Qt::DefaultLocaleLongDate)));
         } else {
             moviePlayedItem->addToolTip(QString(tr("played once, last time on %1")).
-                    arg(QDateTime::fromMSecsSinceEpoch(static_cast<qint64>(timeStamp)).toString(Qt::DefaultLocaleLongDate)));
+                    arg(QDateTime::fromMSecsSinceEpoch(timeStamp).toString(Qt::DefaultLocaleLongDate)));
         }
     }
 }
